scripts/requirements.txt
slack-sdk>=3.26.0
anthropic>=0.25.0
python-dotenv>=1.0.0
